file_util: move PathError and its helper trait from repo module
It's generally useful in order to attach context to io::Error.
