+++ lib/src/file_util.rs
@@ -13,10 +13,32 @@
 // limitations under the License.
 
 use std::fs::File;
-use std::iter;
 use std::path::{Component, Path, PathBuf};
+use std::{io, iter};
 
 use tempfile::{NamedTempFile, PersistError};
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+#[error("Cannot access {path}")]
+pub struct PathError {
+    pub path: PathBuf,
+    #[source]
+    pub error: io::Error,
+}
+
+pub(crate) trait IoResultExt<T> {
+    fn context(self, path: impl AsRef<Path>) -> Result<T, PathError>;
+}
+
+impl<T> IoResultExt<T> for io::Result<T> {
+    fn context(self, path: impl AsRef<Path>) -> Result<T, PathError> {
+        self.map_err(|error| PathError {
+            path: path.as_ref().to_path_buf(),
+            error,
+        })
+    }
+}
 
 /// Turns the given `to` path into relative path starting from the `from` path.
 ///
